requirement.txt
Django==4.1
django-environ==0.9.0
Pillow==9.2.0

